+++ controllers/contacts/index.js
@@ -1,0 +1,13 @@
+const getAll = require("./getAll");
+const getById = require("./getById");
+const addContact = require("./addContact");
+const removeContact = require("./removeContact");
+const updateContact = require("./updateContact");
+
+module.exports = {
+  getAll,
+  getById,
+  addContact,
+  removeContact,
+  updateContact,
+};
